WebStorage/tests/approved/storage_local_key_null.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_key_null</h3>
	</p>
	
	<p>
		<h3>
			<div id="result" >
				<span id="resultMsg"><span class="fail">FAIL<span></span>
			</div>
		</h3>
	</p>

	<hr>

	<div id="log"></div>

	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();

			localStorage.setItem("name", "user1");
			localStorage.setItem("age", "20");
			localStorage.setItem("a", "1");
			localStorage.setItem("b", "2");

			try {

				test(function() {
					assert_equals(localStorage.key(-1), null, "localStorage.key(-1)")
				}, "key method test 1")

				test(function() {
					assert_equals(localStorage.key(4), null, "localStorage.key(4)")
				}, "key method test 2")
				
			} catch (e) {
				test(function()
				{
					assert_unreached("key method is failed.");
				}, "key method test")
			}
		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
